Added student controller

+++ db/seed.sql
@@ -5,7 +5,7 @@ id SERIAL PRIMARY KEY,
 first_name TEXT NOT NULL,
 last_name TEXT NOT NULL,
 home_state TEXT,
-GPA INT
+GPA NUMERIC
 );
 
 CREATE TABLE enrollment_table(
